transfer methods from controller to service
